Refactor: Relocate data, network, and repository files
This commit relocates files related to data models, network services, and repositories into more specific subpackages within the `data` package.
